project_data_science/requirements.txt
selenium>=4.14.0
webdriver-manager>=3.8.6
python-dotenv>=1.0.1
PyJWT>=2.10.1